requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
pydantic-settings
playwright
beautifulsoup4
email-validator